File/requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tldextract>=5.1.2
cssselect>=1.2.0
reportlab>=4.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.12.0
PyJWT>=2.8.0
